bugfix
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@5741 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/media/plugins/wav_reader/WavReaderPlugin.cpp
@@ -140,7 +140,7 @@ WavReader::Sniff(int32 *streamCount)
 				}
 				break;
 			default:
-				TRACE("WavReader::Sniff: ignoring chunk 0x%08lx of %lu bytes\n", UINT32(chunk.fourcc), chunk.len);
+				TRACE("WavReader::Sniff: ignoring chunk 0x%08lx of %lu bytes\n", UINT32(chunk.fourcc), UINT32(chunk.len));
 				break;
 		}
 		pos += UINT32(chunk.len);
